Closes #3
The array assign to the autocomplete was not being updated whenever a
product/provider contain in it was update or deleted. This lead in still
showing deleted product/providers as suggestions or showing old names
for both on the suggestion

This was fixed by assigning the array directly to the filter function
and updating and removing respectively from array whenver user updated a
field or deleted one of the items.